+++ inst/yupana/pkgs.R
@@ -4,26 +4,19 @@
 #> open https://flavjack.github.io/inti/
 #> open https://flavjack.shinyapps.io/yupanapro/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2021-03-24
+#> date .: 2021-04-08
 # -------------------------------------------------------------------------
 
 suppressPackageStartupMessages({
-  
   library(inti)
   library(bslib)
   library(metathis)
   library(googlesheets4)
   library(googleAuthR)
   library(shinydashboard)
-  library(ggpubr)
+  library(cowplot)
   library(corrplot)
   library(tidyverse)
   library(FactoMineR)
-  
 })
-
-
-
-
-