+++ inst/yupana/setup.R
@@ -4,7 +4,7 @@
 #> open https://flavjack.github.io/inti/
 #> open https://flavjack.shinyapps.io/yupanapro/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2021-03-24
+#> date .: 2021-04-08
 # -------------------------------------------------------------------------
 
 cran <- c(
@@ -14,7 +14,7 @@ cran <- c(
   , "googlesheets4"
   , "googleAuthR"
   , "shinydashboard"
-  , "ggpubr"
+  , "cowplot"
   , "corrplot"
   , "tidyverse"
   , "FactoMineR"