+++ R/fieldbook_summary.R
@@ -12,8 +12,8 @@
 #'
 #' @details
 #'
-#' For compare the factors you should use ":". For example, to compare
-#' treatment1 and treatment2: \code{treatment1:treatment2}.
+#' For compare the factors you should use "*". For example, to compare
+#' treatment1 and treatment2: \code{treatment1*treatment2}.
 #'
 #' @return data frame
 #'